requirements.txt
aiofiles==0.8.0
aioredis==2.0.1
alembic==1.8.1
async-timeout==4.0.2
asyncpg==0.26.0
certifi==2022.6.15
gino==1.0.1
h11==0.11.0
hiredis==2.0.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-editor==1.0.4
redis==4.3.4
rfc3986==1.5.0
sanic==22.6.2
Sanic-Cors==2.1.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
ujson==5.4.0
uvloop==0.16.0
websockets==10.3
pytz==2022.2.1
bcrypt==3.2.2
APScheduler
pycryptodome
pandas
openpyxl
xlrd
PyYAML==6.0
pytablewriter==0.64.2
dacite~=1.6.0
pydantic==1.10.2
jsonref==0.2
argparse==1.4.0